requirements.txt
flake8
pylint
pytest
selenium
webdriver-manager
python-dotenv
allure-pytest
